# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pedeve App Backend API dengan struktur Clean Architecture",
        "title": "Pedeve App Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/shareholder-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar semua jenis pemegang saham yang aktif. User reguler hanya melihat jenis pemegang saham aktif, superadmin/administrator bisa melihat semua termasuk yang tidak aktif.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company Management"
                ],
                "summary": "Ambil Semua Shareholder Types",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include inactive shareholder types (hanya superadmin/administrator)",
                        "name": "include_inactive",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar jenis pemegang saham berhasil diambil",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.ShareholderTypeModel"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat jenis pemegang saham baru. Hanya superadmin dan administrator yang dapat membuat jenis pemegang saham baru. Jika jenis pemegang saham dengan nama yang sama sudah ada (tidak aktif), akan diaktifkan kembali.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company Management"
                ],
                "summary": "Buat Shareholder Type Baru",
                "parameters": [
                    {
                        "description": "Shareholder type data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Jenis pemegang saham berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/domain.ShareholderTypeModel"
                        }
                    },
                    "400": {
                        "description": "Invalid request atau jenis pemegang saham sudah ada",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (hanya superadmin/administrator)",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/shareholder-types/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengupdate jenis pemegang saham. Hanya superadmin dan administrator yang dapat mengupdate jenis pemegang saham.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company Management"
                ],
                "summary": "Update Shareholder Type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shareholder Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Shareholder type data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jenis pemegang saham berhasil diupdate",
                        "schema": {
                            "$ref": "#/definitions/domain.ShareholderTypeModel"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (hanya superadmin/administrator)",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus jenis pemegang saham. Hanya superadmin yang dapat menghapus secara permanen. Administrator hanya dapat soft delete (menonaktifkan). Jika jenis pemegang saham masih digunakan, akan dilakukan soft delete.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company Management"
                ],
                "summary": "Hapus Shareholder Type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shareholder Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jenis pemegang saham berhasil dihapus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request atau jenis pemegang saham masih digunakan",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (hanya superadmin/administrator)",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.ShareholderTypeModel": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "description": "User yang membuat",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "description": "Soft delete: false jika dihapus",
                    "type": "boolean"
                },
                "name": {
                    "description": "Nama jenis pemegang saham (unique)",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_count": {
                    "description": "Jumlah pemegang saham yang menggunakan jenis ini",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Atau menggunakan httpOnly cookie (auth_token) untuk keamanan yang lebih baik.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}